GIX-1777: ExpandableSection working
